Extract init_locks.hpp.
