+++ stl/src/syncstream.cpp
@@ -11,6 +11,8 @@
 #include <shared_mutex>
 #include <utility>
 
+#include "init_locks.hpp"
+
 #pragma warning(disable : 4074)
 #pragma init_seg(compiler)
 static std::_Init_locks initlocks;
